# Faster read using Cython code