--- a/news/cython-reader.rst
+++ b/news/cython-reader.rst
@@ -1,0 +1,32 @@
+New features
+------------
+
+* Reading typical numeric data now has less overhead, as it can use a fast code
+  path implemented mostly in Cython. Making many small reads from the same
+  dataset can be as much as 10 times faster, but there are many factors that
+  can affect performance.
+
+Deprecations
+------------
+
+* <news item>
+
+Exposing HDF5 functions
+-----------------------
+
+* <news item>
+
+Bug fixes
+---------
+
+* <news item>
+
+Building h5py
+-------------
+
+* <news item>
+
+Development
+-----------
+
+* <news item>