@takluyver 2019-11-14
I think this is ready for initial review. I've plumbed it into the high-level reading code, and all the tests are passing.