@scopatz 2019-11-14
This looks awesome! Can you add a news entry, please?