@kif 2019-11-18
I got 3x speed up in reading. Congratulations.